verification/server/pyproject.toml
[project]
name = "iden-verification-server"
version = "0.1.0"
description = "IDEN verification server — /v1/capture + /v1/verify (RetinaFace + ArcFace via InsightFace, pgvector)"
requires-python = ">=3.13,<3.14"
dependencies = [
    "fastapi[standard]>=0.136.0",
    "pydantic>=2.13.2",
    "pydantic-settings>=2.7.0",
    "sqlalchemy[asyncio]>=2.0.49",
    "asyncpg>=0.31.0",
    "pgvector>=0.4.0",
    "insightface>=0.7.3",
    "onnxruntime-gpu>=1.18.0",
    "opencv-python-headless>=4.9.0",
    "numpy>=1.26",
    "nvidia-cublas-cu12>=12.9.2.10",
    "nvidia-cudnn-cu12>=9.21.0.82",
    "nvidia-cuda-runtime-cu12>=12.9.79",
    "greenlet>=3.4.0",
    "nvidia-cufft-cu12>=11.4.1.4",
    "nvidia-curand-cu12>=10.3.10.19",
    "nvidia-cusparse-cu12>=12.5.10.65",
]
